+++ docs/history.rst
@@ -1,8 +1,8 @@
 History
 =======
 
-Latest
-------
+0.10.0
+-------
 - DEP: Drop Python 3.7 support (issue #451)
 - ENH: Add GCPs reading and writing (issue #376)